# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/users/profile/edited:
    put:
      tags:
      - rest-user
      summary: Edit user profile
      operationId: editUser
      parameters:
      - name: username
        in: query
        required: true
        schema:
          type: string
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: surname
        in: query
        required: true
        schema:
          type: string
      - name: encodedPassword
        in: query
        required: true
        schema:
          type: string
      - name: confirmEncodedPassword
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: Changes made successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User_UserDetails'
        "403":
          description: Only access to registered user
...
  /api/users/new:
    post:
      tags:
      - rest-user
      summary: Create a new user
      operationId: createUser
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      - name: username
        in: query
        required: true
        schema:
          type: string
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: surname
        in: query
        required: true
        schema:
          type: string
      - name: encodedPassword
        in: query
        required: true
        schema:
          type: string
      - name: confirmEncodedPassword
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "201":
          description: Account created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User_UserDetails'
        "400":
          description: There is already a user with this email or username
        "403":
          description: Only unregistered users can create a new one
...
  /api/users/profile:
    get:
      tags:
      - rest-user
      summary: Get user profile by the email
      operationId: getUser
      responses:
        "404":
          description: Not Found
        "200":
          description: Found the user profile
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User_UserDetails'
        "403":
          description: Only access to registered user
  /api/users/profile/treeHeight:
    get:
      tags:
      - rest-user
      summary: Show the user progress with the tree height
      operationId: getHeight
      responses:
        "404":
          description: Not Found
        "403":
          description: Only access to registered user
        "200":
          description: List of all categories with their height
          content:
            '*/*':
              schema:
                type: string
  /api/users/profile/finishedPlans:
    get:
      tags:
      - rest-user
      summary: Show the user progress with the number of finished plans per category
      operationId: getFinishedPlans
      responses:
        "404":
          description: Not Found
        "200":
          description: List of all categories with their number of finished plans
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Only access to registered user
  /api/users/profile/favPlans:
    get:
      tags:
      - rest-user
      summary: Show the user progress with the number of favourites plans per category
      operationId: getFavPlans
      responses:
        "404":
          description: Not Found
        "200":
          description: List of all categories with their number of favourite plans
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Only access to registered user
  /api/users/completedPlans:
    get:
      tags:
      - rest-user
      summary: Show completed plans by email
      operationId: getCompletedTasksByUser
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: User not found
          content: {}
        "200":
          description: List of all plans completed by the user
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Only access to admin account
  /api/users/completedPlans/:
    get:
      tags:
      - rest-user
      summary: Show all completed plans by users
      operationId: getCompletedTasks
      responses:
        "404":
          description: Not Found
        "403":
          description: Only access to admin account
        "200":
          description: List of all plans completed by all the users
          content:
            '*/*':
              schema:
                type: string
...
components:
  schemas:
    Category_UserDetails:
      type: object
      properties:
        userFavoritesCategory:
          type: array
          items:
            $ref: '#/components/schemas/User_UserDetails'
        description:
          type: string
    Tree_UserDetails:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User_UserDetails'
        category:
          $ref: '#/components/schemas/Category_UserDetails'
        last_update:
          type: integer
          format: int64
        date:
          type: string
        imagePath:
          type: string
    User_UserDetails:
      type: object
      properties:
        trees:
          type: array
          items:
            $ref: '#/components/schemas/Tree_UserDetails'
        email:
          type: string
        username:
          type: string
        name:
          type: string
        surname:
          type: string
        encodedPassword:
          type: string
        confirmEncodedPassword:
          type: string
        roles:
          type: array
          items:
            type: string
        password:
          type: string
          writeOnly: true
...